requirements.txt
openai
pyyaml
python-dotenv
termcolor
